Add cell style back
